login fixed + database
 Git Push Notes:

1. Fixed an issue that was preventing the application from running properly.
   - Updated the project's POM file to match the example.
   - Added a missing bean in the security configuration to resolve a bean-related problem for the mapper model.
   - Corrected a typo from "findbyzipcode" to "findbyzipCode" in the venue repository.

2. Database changes:
   - Added SQL scripts to populate data in the database.
   - Updated the README to include the database schema.
   - Added "contact" and "website" attributes to both vendor and venue entities.

3. Addressed a login issue.
   - Made changes to the authentication logic to resolve the login problem.
   - Modified only one role and added a role prefix to role names for consistency. "role_"

+++ src/main/java/com/saiyingge/localVenue/dto/ReviewDTO.java
@@ -13,12 +13,12 @@ import java.util.List;
 @NoArgsConstructor
 @ToString
 public class ReviewDTO {
-    private Long id;
+    private int id;
     private String content;
     private Integer rating;
 
-    private Long userId;
-    private Long venueId;
-    private Long vendorId;
+    private int userId;
+    private int venueId;
+    private int vendorId;
     private String date;
 }
